+++ app/Http/Controllers/Controller.php
@@ -2,27 +2,17 @@
 
 namespace App\Http\Controllers;
 
-use App\Preference;
-use App\ThConcept;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller as BaseController;
-use Illuminate\Support\Facades\View;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 
     public function __construct() {
-        $preferences = Preference::all();
-        $preferenceValues = [];
-        foreach($preferences as $p) {
-            $preferenceValues[$p->label] = Preference::decodePreference($p->label, json_decode($p->default_value));
-        }
-
-        View::share('p', $preferenceValues);
     }
 
     public function hasInput(Request $request) {